requirements.txt
cffi==1.17.1
ema-pytorch==0.7.7
fcd-torch==1.0.7
gmpy2==2.2.1
hydra-core==1.3.2
loguru==0.7.3
matplotlib==3.7.1
munkres==1.1.4
overrides==7.3.1
pandas==1.4.0
pip==24.2
pyemd==1.0.0
PyGObject==3.50.0
PyGSP==0.5.1
reportlab==4.4.1
rlPyCairo==0.2.0
scikit-learn==1.7.0
SQLAlchemy==2.0.41
torch-geometric==2.6.1
torch-scatter==2.1.2
torchmetrics==1.6.3
tzdata==2025.2
unicodedata2==16.0.0
wandb==0.19.1
wheel==0.45.1
zstandard==0.23.0
